+++ lib/fun_with/files/core_extensions/array.rb
@@ -1,5 +1,12 @@
-class Array
-  def fwf_blank?
-    self.length == 0
+module FunWith
+  module Files
+    module CoreExtensions
+      module Array
+        def fwf_blank?
+          self.length == 0
+        end
+      end
+    end
   end
 end
+