+++ lib/fun_with/files/core_extensions/false.rb
@@ -1,5 +1,11 @@
-class FalseClass
-  def fwf_blank?
-    true
+module FunWith
+  module Files
+    module CoreExtensions
+      module FalseClass
+        def fwf_blank?
+          true
+        end
+      end
+    end
   end
 end