+++ lib/fun_with/files/core_extensions/nil.rb
@@ -1,5 +1,12 @@
-class NilClass
-  def fwf_blank?
-    true
+module FunWith
+  module Files
+    module CoreExtensions
+      module NilClass
+        def fwf_blank?
+          true
+        end
+      end
+    end
   end
 end
+