+++ lib/fun_with/files/core_extensions/object.rb
@@ -1,9 +1,15 @@
-class Object
-  def fwf_blank?
-    false
-  end
+module FunWith
+  module Files
+    module CoreExtensions
+      module Object
+        def fwf_blank?
+          false
+        end
   
-  def fwf_present?
-    ! self.fwf_blank?
+        def fwf_present?
+          ! self.fwf_blank?
+        end
+      end
+    end
   end
 end