+++ test/helper.rb
@@ -1,6 +1,3 @@
-# require 'rubygems'
-# require 'bundler'
-# require 'debugger'
 require 'fun_with_testing'
 
 
@@ -35,10 +32,23 @@ class FunWith::Files::TestCase < FunWith::Testing::TestCase
   end
   
   def empty_temp_directory
-    tmp = FunWith::Files.root("test", "tmp")
+    tmp = FunWith::Files.root( "test", "tmp" )
     tmp.empty!
     assert_directory tmp
     puts tmp.glob(:all)
     assert_empty_directory tmp
+  end
+  
+  def if_internet_works( &block )
+    `ping -c 1 google.com 2>&1 >> /dev/null`  # TODO: Portability issue
+    connection_detected = $?.success?
+      
+    if block_given?
+      if connection_detected   # TODO:  How to tell difference between "no internet" and "no ping utility?"
+        yield
+      else
+        puts "No internet connection.  Skipping."
+      end
+    end
   end
 end